[feature] #1970: Add optional schema endpoint
