+++ cli/Cargo.toml
@@ -17,8 +17,10 @@ dex = ["iroha_core/dex"]
 roles = ["iroha_core/roles", "iroha_permissions_validators/roles"]
 telemetry = ["iroha_telemetry", "iroha_core/telemetry"]
 dev-telemetry = ["iroha_core/dev-telemetry", "iroha_telemetry"]
-default = ["bridge", "telemetry"]
+schema-endpoint = ["iroha_schema_bin"]
 test-network = []
+
+default = ["bridge", "telemetry", "schema-endpoint"]
 
 [badges]
 is-it-maintained-issue-resolution = { repository = "https://github.com/hyperledger/iroha" }
@@ -37,6 +39,7 @@ iroha_version = { version = "=2.0.0-pre-rc.2", path = "../version", features = [
 iroha_config = { version = "=2.0.0-pre-rc.2", path = "../config" }
 iroha_crypto = { version = "=2.0.0-pre-rc.2", path = "../crypto" }
 iroha_p2p = { version = "=2.0.0-pre-rc.2", path = "../p2p" }
+iroha_schema_bin = { version = "=2.0.0-pre-rc.2", path = "../schema/bin", optional = true }
 
 eyre = "0.6.5"
 futures = { version = "0.3.17", default-features = false, features = ["std", "async-await"] }
